+++ cmd/searchgirl/main.go
@@ -8,7 +8,7 @@ import (
 	"os"
 )
 
-const version = "0.5.2"
+const version = "0.6.0"
 
 func main() {
 	// No arguments = serve. "Fácil de usar" starts here.